azync: clean up awaiters once select resolves
select would leave some dangling references in the `_awaiter`s in the
select's arguments leading to the following resume after return
situation:

```
local a
local b

-- this frame is selecting on a and b
async(function()
	local result = selectAwait(
		async(function() a = currentFrame() suspend() return 1 end),
		async(function() b = currentFrame() suspend() return 2 end)
	)

	print(result)
end)

resume(a) -- a resolves, which makes the selecting frame also resolve

-- but if b still has the selecting frame as its awaiter
resume(b) -- this resumes a function that already returned
```

before returning, now we set the `_awaiter`s field to nil so this can't
happen any more.

This does raise the question of: should a frame that has been
`select`ed, be `await`able after the select resolves. In a more dynamic
language like Lua, this seems fine. For Zig, the `select` proposal has
been held up for many reasons so it is hard to learn from them without a
corresponding implementation :P